backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pydantic==1.10.5
pytz==2022.7.1
requests==2.28.2
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
starlette==0.25.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.20.0
